requirements.txt
# LSED - LLM-enhanced Social Event Detection
# Requirements file

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# PyTorch (adjust version based on CUDA availability)
torch>=2.0.0

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.20.0

# Optional: spaCy for Word2Vec (alternative vectorizer)
# spacy>=3.0.0
# en-core-web-sm model: python -m spacy download en_core_web_sm

# LLM interaction
requests>=2.28.0

# Configuration
pyyaml>=6.0

# Progress bar
tqdm>=4.64.0

# Logging and utilities
colorlog>=6.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.990

# Visualization (optional)
# matplotlib>=3.5.0
# seaborn>=0.12.0

# Jupyter (optional)
# jupyter>=1.0.0
# ipywidgets>=8.0.0
